refactor(mailbridge-tests): split MailBridgeProgramTests under 500-line cap
- Split MailBridgeProgramTests.cs (573) into three partials: base 264,
  RunAsync 268, SendMail 71 — all under the 500-line limit, no test
  removed or weakened (587 passing, Integration excluded)
- Correct acceptance-criteria-map.md file-size claim to match the split
- Record live-COM integration exception dossier (no live-Outlook host;
  AC-06 / AC-10 a,b covered-by-design)
- Add issue-75 feature-review round-1 audit artifacts

Refs: #75
