+++ docs/features/active/hostadapter-sendmail-com-send-75/evidence/other/acceptance-criteria-map.md
@@ -19,14 +19,17 @@ acceptance-criteria-tracking skill, status is tracked here rather than by reform
 | AC-08 | `saveToSentItems` defaults true; maps to `DeleteAfterSubmit = !saveToSentItems` | P4-T2, P5-T1, P5-T4, P8-T2 | `Send_mail_save_to_sent_items_should_default_to_true_when_absent`; `MailContractsTests` default-true; `BuildSendMail` default true; OutlookComMailSender sets `DeleteAfterSubmit = !SaveToSentItems` | PASS |
 | AC-09 | Send-on-behalf deferred to PI-1; seam accepts future `fromEmailAddress` without breaking callers; documented | P2-T2, P4-T1, P10-T2 | `SendMailComRequest.FromEmailAddress` optional param; `IHostAdapterClient.SendMailAsync` XML doc PI-1 note; README PI-1 note | PASS |
 | AC-10 | Tests cover integration Sent Items + COM path; endpoint unit test with mocked runner; Core client POST/body/token tests; RPC dispatch fake-sender tests; contract-coverage test for send_mail | P1-T2, P5-T3, P5-T4, P7-T3, P8-T5, P8-T6, P9-T1 | `FakeOutlookMailSender`; all test files above | PASS |
-| AC-11 | Seven-stage toolchain passes; line>=85%, branch>=75%; no regression on changed lines; only documented `[ExcludeFromCodeCoverage]` (integration-covered); no file > 500 lines; additive contracts only | P11-T1..P11-T6 | final-format/lint/typecheck/architecture/test-coverage/coverage-delta artifacts; line 90.25% / branch 79.35%; no file > 500; additive only | PASS |
+| AC-11 | Seven-stage toolchain passes; line>=85%, branch>=75%; no regression on changed lines; only documented `[ExcludeFromCodeCoverage]` (integration-covered); no file > 500 lines; additive contracts only | P11-T1..P11-T6; remediation R-1 split (2026-06-16) | final-format/lint/typecheck/architecture/test-coverage/coverage-delta artifacts; line 90.25% / branch 79.35%; after the R-1 remediation split every changed `.cs` file is <= 500 lines (largest changed test file `MailBridgeProgramTests.RunAsync.cs` 268; largest changed file overall `OutlookScanner.cs` 465 — see `evidence/qa-gates/remediation-file-sizes.md`); additive only | PASS |
 
 ## Suppression / file-size / additive-contract confirmation
 
 - New suppressions: only `[ExcludeFromCodeCoverage]` on 3 live-COM-only members of `OutlookComMailSender`,
   each annotated as integration-test-covered. No new analyzer/nullable suppressions.
-- No file exceeds 500 lines (largest touched: OutlookScanner.cs 465; MailBridgeRuntimeTestDoubles.cs 495;
-  PipeRpcWorker.cs 438; Program.cs 436).
+- No file exceeds 500 lines. The initial branch left `tests/OpenClaw.MailBridge.Tests/MailBridgeProgramTests.cs`
+  at 573 lines; the R-1 remediation split (2026-06-16) divided it into three behavior-preserving partial-class
+  files (`MailBridgeProgramTests.cs` 264, `MailBridgeProgramTests.RunAsync.cs` 268, `MailBridgeProgramTests.SendMail.cs` 71),
+  so every changed `.cs` file is now <= 500 lines. Largest changed files after the split: OutlookScanner.cs 465;
+  PipeRpcWorker.cs 438; Program.cs 436. Verified by `evidence/qa-gates/remediation-file-sizes.md`.
 - All contract changes additive: BridgeMethods.SendMail added to All; new SendMail* DTOs; new
   IHostAdapterClient.SendMailAsync; new route; new RPC verb. No existing member/route/verb altered or
   removed. No major version bump.
