+++ docs/features/active/hostadapter-sendmail-com-send-75/evidence/qa-gates/remediation-coverage-delta.md
@@ -1,0 +1,21 @@
+# Remediation Final QA — Coverage No-Regression Delta
+
+Timestamp: 2026-06-16T08-06
+Command: comparison of baseline `evidence/remediation-baseline/test-coverage.md` (P0-T3) vs post-change `evidence/qa-gates/remediation-final-test-coverage.md` (P4-T5)
+EXIT_CODE: 0
+
+Output Summary:
+| Metric | Baseline (P0-T3) | Post-change (P4-T5) | Delta | Threshold | Verdict |
+|---|---|---|---|---|---|
+| Line coverage (combined) | 90.25% (4028/4463) | 90.25% (4028/4463) | 0.00% | >= 85% | PASS |
+| Branch coverage (combined) | 79.36% (911/1148) | 79.36% (911/1148) | 0.00% | >= 75% | PASS |
+| Passing tests (Integration excluded) | 587 | 587 | 0 | no regression | PASS |
+
+Changed/new-code coverage: the R-1 remediation changed only test source (split of `MailBridgeProgramTests.cs`
+into three behavior-preserving partial-class files) and documentation/evidence. No production (`src/**`) lines
+were added, modified, or removed, so there are no changed production lines whose coverage could regress. Test
+code is excluded from the coverage surface per `.claude/rules/general-unit-test.md`, which is why per-assembly
+and combined coverage values are bit-for-bit identical to the baseline.
+
+No-regression verdict: PASS. Combined line >= 85% and branch >= 75% both hold; no changed-line coverage was
+reduced (no production lines changed); test count did not regress.
