+++ docs/features/active/hostadapter-sendmail-com-send-75/evidence/qa-gates/remediation-final-architecture.md
@@ -1,0 +1,6 @@
+# Remediation Final QA — Architecture Boundaries
+
+Timestamp: 2026-06-16T08-05
+Command: `dotnet build OpenClaw.MailBridge.sln -c Debug` (plus `git diff --name-only 0cb7de6..HEAD | grep .csproj` and `git status` csproj check)
+EXIT_CODE: 0
+Output Summary: Build succeeded, 0 Warning(s) / 0 Error(s). No `.csproj` files were changed by this remediation (no tracked diff and no untracked/modified csproj), so the project-reference graph in `.claude/rules/architecture-boundaries.md` is unchanged. The remediation touched only test source files (the `MailBridgeProgramTests` partial split) and documentation/evidence. COM confinement to `OpenClaw.MailBridge` is preserved; the split moved client-program unit tests within the existing `OpenClaw.MailBridge.Tests` project only. No new project references, no circular references introduced.
