+++ docs/features/active/hostadapter-sendmail-com-send-75/evidence/qa-gates/remediation-final-format.md
@@ -1,0 +1,6 @@
+# Remediation Final QA — Format
+
+Timestamp: 2026-06-16T08-05
+Command: `csharpier format .` then `csharpier check .`
+EXIT_CODE: 0 (both)
+Output Summary: `csharpier format .` formatted 193 files in 295ms with no content changes to source; `csharpier check .` confirmed 193 files conform (EXIT_CODE 0). Idempotent — no loop restart required.
