+++ docs/features/active/hostadapter-sendmail-com-send-75/evidence/qa-gates/remediation-final-lint.md
@@ -1,0 +1,6 @@
+# Remediation Final QA — Lint / Analyzers
+
+Timestamp: 2026-06-16T08-05
+Command: `dotnet build OpenClaw.MailBridge.sln -c Debug -p:EnableNETAnalyzers=true -p:EnforceCodeStyleInBuild=true`
+EXIT_CODE: 0
+Output Summary: Build succeeded. 0 Warning(s), 0 Error(s). .NET SDK analyzers and code-style enforcement pass with no diagnostics. No new suppressions introduced.
