+++ docs/features/active/hostadapter-sendmail-com-send-75/evidence/qa-gates/remediation-final-test-coverage.md
@@ -1,0 +1,17 @@
+# Remediation Final QA — Test + Coverage
+
+Timestamp: 2026-06-16T08-06
+Command: `dotnet test OpenClaw.MailBridge.sln -c Debug --settings mailbridge.runsettings --collect:"XPlat Code Coverage" --filter "TestCategory!=Integration"`
+EXIT_CODE: 0
+
+Output Summary:
+- Result: Passed. Failed: 0, Skipped: 3, total passing 587 (Integration excluded). No test-count regression vs the 587 baseline.
+  - OpenClaw.HostAdapter.Tests: 100 passed.
+  - OpenClaw.Core.Tests: 210 passed.
+  - OpenClaw.MailBridge.Tests: 277 passed, 3 skipped (same three platform/publish gated skips as baseline).
+- Combined post-change coverage (sum of per-assembly cobertura across the three test projects):
+  - Line: 4028/4463 = 90.25% (>= 85% PASS)
+  - Branch: 911/1148 = 79.36% (>= 75% PASS)
+- Per-project rates (unchanged from baseline): Core.Tests line 89.61% / branch 78.44%; HostAdapter.Tests line 87.70% / branch 67.19%; MailBridge.Tests line 93.08% / branch 86.92%.
+
+Coverage is identical to the Phase 0 baseline because the R-1 split moved test code only (excluded from the coverage surface); no production code was changed. No loop restart required.
