+++ docs/features/active/hostadapter-sendmail-com-send-75/evidence/qa-gates/remediation-final-typecheck.md
@@ -1,0 +1,6 @@
+# Remediation Final QA — Nullable Type-Check
+
+Timestamp: 2026-06-16T08-05
+Command: `dotnet build OpenClaw.MailBridge.sln -c Debug -p:TreatWarningsAsErrors=true`
+EXIT_CODE: 0
+Output Summary: Build succeeded. 0 Warning(s), 0 Error(s). Nullable reference-type flow analysis produces no warnings-as-errors. No `#nullable disable` introduced.
