+++ docs/features/active/hostadapter-sendmail-com-send-75/evidence/qa-gates/remediation-split-loop.md
@@ -1,0 +1,23 @@
+# Remediation QA Gate — Split C# Loop (P1-T4)
+
+Timestamp: 2026-06-16T08-03
+
+## Stage 1 — Format
+Command: `csharpier format .`
+EXIT_CODE: 0
+Output Summary: Formatted 193 files in 407ms. Re-verified idempotent with `csharpier check .` (EXIT_CODE 0, 193 files). The two new partial files were already CSharpier-conformant; no tracked file content changed by formatting.
+
+## Stage 2 — Lint / Analyzers + Code Style
+Command: `dotnet build OpenClaw.MailBridge.sln -c Debug -p:EnableNETAnalyzers=true -p:EnforceCodeStyleInBuild=true`
+EXIT_CODE: 0
+Output Summary: Build succeeded. 0 Warning(s), 0 Error(s). All seven projects compiled.
+
+## Stage 3 — Test + Coverage
+Command: `dotnet test OpenClaw.MailBridge.sln -c Debug --settings mailbridge.runsettings --collect:"XPlat Code Coverage" --filter "TestCategory!=Integration"`
+EXIT_CODE: 0
+Output Summary: Passed. Failed: 0, Skipped: 3, total passing 587 (Integration excluded) — no regression from the 587 baseline.
+- OpenClaw.HostAdapter.Tests: 100 passed.
+- OpenClaw.Core.Tests: 210 passed.
+- OpenClaw.MailBridge.Tests: 277 passed, 3 skipped (same three platform/publish skips as baseline).
+
+Behavior of the split is preserved: the MailBridgeProgramTests suite reports the same passing count as before the split (the three partial files compose into one `partial class MailBridgeProgramTests`). No stage failed or auto-fixed files, so no loop restart was required.
