+++ docs/features/active/hostadapter-sendmail-com-send-75/evidence/remediation-baseline/format.md
@@ -1,0 +1,6 @@
+# Remediation Baseline — Format
+
+Timestamp: 2026-06-16T07-57
+Command: `csharpier check .`
+EXIT_CODE: 0
+Output Summary: Checked 191 files in 432ms. No formatting violations. Baseline format state is clean before the R-1 split.
